third_party/blink/web_tests/platform/generic/fast/events/touch/multi-touch-grouped-targets-expected.txt

Tests that the an event is sent for every touch listener, and target touches contains all the points for that target

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS event.touches.length is 3
PASS event.changedTouches.length is 3
PASS event.touches.length is 3
PASS event.changedTouches.length is 3
PASS event.touches.length is 3
PASS event.changedTouches.length is 3
PASS event.touches.length is 3
PASS event.changedTouches.length is 3
PASS event.type is "touchend"
PASS event.touches.length is 0
PASS event.targetTouches.length is 0
PASS event.changedTouches.length is 3
PASS event.type is "touchend"
PASS event.touches.length is 0
PASS event.targetTouches.length is 0
PASS event.changedTouches.length is 3
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
